Formatting

+++ include/quartz/rpc/payloads/LEDFramebufferSetPayload.hpp
@@ -1,14 +1,14 @@
 #pragma once
+#include "quartz/utils/Color32.hpp"
 #include <array>
 #include <cstdint>
-#include "quartz/utils/Color32.hpp"
 
 namespace quartz::rpc::payloads
 {
     template <std::size_t N>
     struct alignas(4) LEDFramebufferSetPayload
     {
-        const std::size_t MatrixSize = N;
+        const std::uint32_t MatrixSize = N;
         std::array<utils::Color32, N> Framebuffer = {};
     };
 }
